test(cli): split app build html asset tests
